+++ core/src/com/bombbird/terminalcontrol/screens/settingsscreen/customsetting/TrafficFlowScreen.kt
@@ -135,8 +135,8 @@ class TrafficFlowScreen(game: TerminalControl): BasicScreen(game, 5760, 3240) {
         labelStyle.fontColor = Color.WHITE
 
         val headerLabel = Label("Arrival Traffic", labelStyle)
-        headerLabel.width = MainMenuScreen.BUTTON_WIDTH.toFloat()
-        headerLabel.height = MainMenuScreen.BUTTON_HEIGHT.toFloat()
+        headerLabel.width = MainMenuScreen.BUTTON_WIDTH
+        headerLabel.height = MainMenuScreen.BUTTON_HEIGHT
         headerLabel.setPosition(5760 / 2.0f - MainMenuScreen.BUTTON_WIDTH / 2.0f, 3240 * 0.85f)
         headerLabel.setAlignment(Align.center)
         stage.addActor(headerLabel)
